backend/requirements.txt
# FastAPI & Server
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
aiosqlite==0.19.0

# Pydantic
pydantic==2.6.3
pydantic-settings==2.1.0

# File Parsing
python-docx==1.1.0
PyPDF2==3.0.1
markdown==3.5.2

# LLM Integration
dashscope==1.14.1

# HTTP Client
httpx==0.27.0

# CORS & Utilities
python-dotenv==1.0.1
